phonebook-app/requirements.txt
flask==3.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
prometheus-client==0.19.0
pymongo==4.6.1
pytest==7.4.3
pytest-mock==3.12.0
requests==2.31.0
pytest-cov==4.1.0  
pytest-xdist==3.3.1 
